+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## 0.1.0
 
-* TODO: Describe initial release.
+Initial release.
+
+`KeyboardHider` widget with optional `HideMode` to either unfocus the current `FocusScope`, or to use `SystemChannels.textInput` to hide the keyboard.
+
+It also includes the `unfocus` and `hideTextInput` helper methods.